+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed:
-- Updated dependencies
+- Upgraded dependencies
 - Show config flag for invalid cert
+- App documents folder on Windows
+
+### Fixed:
+- App ID on Windows and Linux
 
 ### Added:
 - Save and restore window position on Windows and Linux